Refactor filter code
